feat(tee): add prover binary
