+++ bin/prover/src/main.rs
@@ -1,0 +1,16 @@
+#![doc = include_str!("../README.md")]
+#![doc(issue_tracker_base_url = "https://github.com/base/base/issues/")]
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
+
+mod cli;
+
+#[tokio::main]
+async fn main() {
+    use clap::Parser as _;
+
+    if let Err(err) = cli::Cli::parse().run().await {
+        eprintln!("Error: {err:?}");
+        std::process::exit(1);
+    }
+}
